Refactor track deletion and file management services
- Extract common file deletion utilities to a new storage/deletion module
- Simplify track deletion queue processing with sequential track handling
- Remove DELETING status from TrackStatus enum
- Streamline error handling in media routes and track deletion service
- Improve file deletion retry mechanism and error tracking
